Close files after tests complete.
